% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_atmo_bulk}
\alias{get_atmo_bulk}
\title{Récupérer l'historique complet Atmo (quotidien ou horaire) sur une plage de dates}
\usage{
get_atmo_bulk(
  df_stations,
  polluant_id,
  date_debut,
  date_fin = Sys.Date(),
  hourly = FALSE
)
}
\arguments{
\item{df_stations}{Data.frame de stations (issu de get_list_stations).}

\item{polluant_id}{Vecteur de codes polluants.}

\item{date_debut}{Date de début ("YYYY-MM-DD" ou objet Date).}

\item{date_fin}{Date de fin. Par défaut aujourd'hui.}

\item{hourly}{Si TRUE, boucle jour par jour et renvoie des valeurs horaires.
Si FALSE (défaut), boucle année par année et renvoie des valeurs journalières.}
}
\value{
Un data.frame consolidé.
}
\description{
Récupérer l'historique complet Atmo (quotidien ou horaire) sur une plage de dates
}
