% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_list_stations}
\alias{get_list_stations}
\title{Récupérer la liste des stations Atmo Auvergne-Rhône-Alpes}
\usage{
get_list_stations()
}
\value{
Un data.frame contenant les colonnes id_station, nom_station, date_debut, date_fin, en_service et typologie.
}
\description{
Récupérer la liste des stations Atmo Auvergne-Rhône-Alpes
}
