% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_atmo}
\alias{get_atmo}
\title{Récupérer l'historique complet sur plusieurs années}
\usage{
get_atmo(
  df_stations,
  polluant_id,
  annee_debut,
  annee_fin = as.numeric(format(Sys.Date(), "\%Y"))
)
}
\arguments{
\item{df_stations}{Data.frame de stations (issu de get_list_stations).}

\item{polluant_id}{Vecteur de codes polluants.}

\item{annee_debut}{Année de départ.}

\item{annee_fin}{Année de fin (par défaut année en cours).}
}
\value{
Un data.frame consolidé.
}
\description{
Récupérer l'historique complet sur plusieurs années
}
