% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_atmo_bulk}
\alias{get_atmo_bulk}
\title{Récupérer les données de mesures pour une année précise}
\usage{
get_atmo_bulk(station_input, polluant_ids, year, df_stations_ref = NULL)
}
\arguments{
\item{station_input}{Un vecteur d'ID de stations ou un data.frame issu de get_list_stations.}

\item{polluant_ids}{Vecteur de codes polluants (ex: c("24", "08")).}

\item{year}{Année au format numérique (ex: 2024).}

\item{df_stations_ref}{Data.frame de référence pour joindre les noms de stations.}
}
\value{
Un data.frame des mesures ou NULL si aucune donnée n'est trouvée.
}
\description{
Récupérer les données de mesures pour une année précise
}
